+++ src/myStyleLoader/insertStyle.js
